Players taking damage is now float to allow enemy damage amplification. Added ability for level to amplify enemy health, damage, speed, and range. .

+++ Teotihuacan/Entities/Player.cs
@@ -382,7 +382,7 @@ Weapon Drain: {1 - weaponLevel.CurrentWeaponLevel * WeaponLevelEnergyDrainDecrem
             }
         }
 
-        public bool TakeDamage(int damageToTake)
+        public bool TakeDamage(float damageToTake)
         {
             bool didTakeDamage = false;
